requirements.txt
asgiref==3.2.3
boto==2.49.0
boto3==1.12.6
botocore==1.15.6
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.2
django-froala-editor==3.1.0
django-js-asset==1.2.2
django-modeltranslation==0.14.4
django-rosetta==0.9.4
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
Pillow==7.0.0
polib==1.1.0
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
